+++ sdk/nodejs/types/output.ts
@@ -5,6 +5,11 @@ import * as pulumi from "@pulumi/pulumi";
 import * as inputs from "../types/input";
 import * as outputs from "../types/output";
 
+export interface ComputeClusterVsanDiskGroup {
+    cache?: string;
+    storages?: string[];
+}
+
 export interface DistributedPortGroupVlanRange {
     maxVlan: number;
     minVlan: number;