+++ sdk/python/pulumi_vsphere/get_compute_cluster.py
@@ -6,7 +6,8 @@ import warnings
 import pulumi
 import pulumi.runtime
 from typing import Union
-from . import utilities, tables
+from . import _utilities, _tables
+
 
 class GetComputeClusterResult:
     """
@@ -28,6 +29,8 @@ class GetComputeClusterResult:
         if resource_pool_id and not isinstance(resource_pool_id, str):
             raise TypeError("Expected argument 'resource_pool_id' to be a str")
         __self__.resource_pool_id = resource_pool_id
+
+
 class AwaitableGetComputeClusterResult(GetComputeClusterResult):
     # pylint: disable=using-constant-test
     def __await__(self):
@@ -39,7 +42,8 @@ class AwaitableGetComputeClusterResult(GetComputeClusterResult):
             name=self.name,
             resource_pool_id=self.resource_pool_id)
 
-def get_compute_cluster(datacenter_id=None,name=None,opts=None):
+
+def get_compute_cluster(datacenter_id=None, name=None, opts=None):
     """
     The `ComputeCluster` data source can be used to discover the ID of a
     cluster in vSphere. This is useful to fetch the ID of a cluster that you want
@@ -73,14 +77,12 @@ def get_compute_cluster(datacenter_id=None,name=None,opts=None):
     :param str name: The name or absolute path to the cluster.
     """
     __args__ = dict()
-
-
     __args__['datacenterId'] = datacenter_id
     __args__['name'] = name
     if opts is None:
         opts = pulumi.InvokeOptions()
     if opts.version is None:
-        opts.version = utilities.get_version()
+        opts.version = _utilities.get_version()
     __ret__ = pulumi.runtime.invoke('vsphere:index/getComputeCluster:getComputeCluster', __args__, opts=opts).value
 
     return AwaitableGetComputeClusterResult(