+++ sdk/python/pulumi_vsphere/get_content_library.py
@@ -6,7 +6,8 @@ import warnings
 import pulumi
 import pulumi.runtime
 from typing import Union
-from . import utilities, tables
+from . import _utilities, _tables
+
 
 class GetContentLibraryResult:
     """
@@ -22,6 +23,8 @@ class GetContentLibraryResult:
         if name and not isinstance(name, str):
             raise TypeError("Expected argument 'name' to be a str")
         __self__.name = name
+
+
 class AwaitableGetContentLibraryResult(GetContentLibraryResult):
     # pylint: disable=using-constant-test
     def __await__(self):
@@ -31,7 +34,8 @@ class AwaitableGetContentLibraryResult(GetContentLibraryResult):
             id=self.id,
             name=self.name)
 
-def get_content_library(name=None,opts=None):
+
+def get_content_library(name=None, opts=None):
     """
     The `ContentLibrary` data source can be used to discover the ID of a Content Library.
 
@@ -51,13 +55,11 @@ def get_content_library(name=None,opts=None):
     :param str name: The name of the Content Library.
     """
     __args__ = dict()
-
-
     __args__['name'] = name
     if opts is None:
         opts = pulumi.InvokeOptions()
     if opts.version is None:
-        opts.version = utilities.get_version()
+        opts.version = _utilities.get_version()
     __ret__ = pulumi.runtime.invoke('vsphere:index/getContentLibrary:getContentLibrary', __args__, opts=opts).value
 
     return AwaitableGetContentLibraryResult(