+++ sdk/python/pulumi_vsphere/get_custom_attribute.py
@@ -6,7 +6,8 @@ import warnings
 import pulumi
 import pulumi.runtime
 from typing import Union
-from . import utilities, tables
+from . import _utilities, _tables
+
 
 class GetCustomAttributeResult:
     """
@@ -25,6 +26,8 @@ class GetCustomAttributeResult:
         if name and not isinstance(name, str):
             raise TypeError("Expected argument 'name' to be a str")
         __self__.name = name
+
+
 class AwaitableGetCustomAttributeResult(GetCustomAttributeResult):
     # pylint: disable=using-constant-test
     def __await__(self):
@@ -35,7 +38,8 @@ class AwaitableGetCustomAttributeResult(GetCustomAttributeResult):
             managed_object_type=self.managed_object_type,
             name=self.name)
 
-def get_custom_attribute(name=None,opts=None):
+
+def get_custom_attribute(name=None, opts=None):
     """
     The `CustomAttribute` data source can be used to reference custom
     attributes that are not managed by this provider. Its attributes are exactly the
@@ -59,13 +63,11 @@ def get_custom_attribute(name=None,opts=None):
     :param str name: The name of the custom attribute.
     """
     __args__ = dict()
-
-
     __args__['name'] = name
     if opts is None:
         opts = pulumi.InvokeOptions()
     if opts.version is None:
-        opts.version = utilities.get_version()
+        opts.version = _utilities.get_version()
     __ret__ = pulumi.runtime.invoke('vsphere:index/getCustomAttribute:getCustomAttribute', __args__, opts=opts).value
 
     return AwaitableGetCustomAttributeResult(